requirements.txt
bcrypt==3.2.2
blinker==1.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
pycparser==2.21
requests==2.27.1
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
